Extract router test utilities into dedicated file
Decouple the two different testing utils as importing from store
initializes API, which we want to avoid in some cases.
This can be refactored in a nicer way later on.
